+++ resources/assets/js/components/ExampleComponent.vue
@@ -1,12 +1,12 @@
 <template>
     <div class="container">
-        <div class="row">
-            <div class="col-md-8 col-md-offset-2">
-                <div class="panel panel-default">
-                    <div class="panel-heading">Example Component</div>
+        <div class="row justify-content-center">
+            <div class="col-md-8">
+                <div class="card card-default">
+                    <div class="card-header">Example Component</div>
 
-                    <div class="panel-body">
-                        I'm an example component!
+                    <div class="card-body">
+                        I'm an example component.
                     </div>
                 </div>
             </div>